pyproject.toml
[project]
name = "multi-agent-research-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.30.0",
    "tavily-python>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pandas>=2.0.0",
    "tiktoken>=0.6.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
    "pydantic>=2.5.0",
    "html5lib>=1.1",
    "langchain-google-genai>=4.2.2",
    "streamlit>=1.56.0",
]
